+++ CHANGELOG.md
@@ -10,6 +10,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Added
 - Enzyme diffusion simulation.
 - Settings export to `gpseq_fromfish`.
+- `gpseq_anim`
+    + `--do-all/-y` option.
+    + `-M`, `-m`, `--labeled` and `--compressed` to export/import masks.
 
 ### Changed
 - Moved distance functions to separate `tools` sub-module.